% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_depth_tib}
\alias{allele_depth_tib}
\title{mpileup and ANNOVAR data merged}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 190 rows and 23 columns.}
\usage{
allele_depth_tib
}
\description{
A tibble of both mpileup and ANNOVAr data - thus it holds the number of
    reads at each mutations
}
\keyword{datasets}
