% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annovar_tib}
\alias{annovar_tib}
\title{ANNOVAR results}
\format{a tibble (411 x 19)
\describe{
    \item{file_id}{file ID}
    \item{Chr}{chromosome}
    \item{Start}{start of variant}
    \item{End}{end of variant}
    \item{Ref}{reference allele}
    \item{Alt}{alternate allele}
    \item{Func.refGene}{where the alteration is}
    \item{Gene.refGene}{gene name}
    \item{GeneDetail.refGene}{}
    \item{ExonicFunc.refGene}{type of mutation}
    \item{AAChange.refGene}{amino acid change}
    \item{aa_mod}{more useable amino acid change}
    \item{file_name}{name of the file}
    \item{data_category}{category of data (are all "Sequencing Reads")}
    \item{data_type}{type of data (are all "Aligned Reads")}
    \item{project_id}{ID of the source project (TCGA-COAD or -READ)}
    \item{case_id}{case_id}
    \item{sample_id}{sample ID}
    \item{sample_type}{from where the sample was taken}
}}
\usage{
annovar_tib
}
\description{
A tibble of the annotation results from ANNOVAR. It was joined with the
    sample information. Get more information on the returned values from
    the \href{http://annovar.openbioinformatics.org/en/latest/}{ANNOVAR documentation}
}
\keyword{datasets}
