% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{choose_one_filename}
\alias{choose_one_filename}
\title{Select the best filename to keep}
\usage{
choose_one_filename(fn)
}
\arguments{
\item{fn}{vector of filenames}
}
\value{
a single filename from the input vector that should be retained
}
\description{
Some cases the sample sheet have multiple files. This function 
    uses a set of (fairly arbitrary) checks to select one filename
}
\examples{
test_fns <- c("TCGA-A6-2672-01B-03D-2301-10_Illumina_gdc_realn.bam",
              "C982.TCGA-A6-2672-01B-03D-2298-08.2_gdc_realn.bam")
choose_one_filename(test_fns)

}
