% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv_tib}
\alias{cnv_tib}
\title{CNA information from TCGA}
\format{a tibble (2539963 x 6)
\describe{
    \item{Sample}{sample ID}
    \item{Chromosome}{chromosome of the CNA}
    \item{Start}{start of the CNA}
    \item{End}{end of the CNA}
    \item{Num_Probes}{number of probes over the region}
    \item{Segment_Mean}{the CNA value equivalent to log2(copy-number/ 2)}
}}
\usage{
cnv_tib
}
\description{
This tibble contains the CNV information on all samples in
    TCGA. It was downloaded from
    \url{https://gdc.cancer.gov/about-data/publications/pancanatlas}.
}
\keyword{datasets}
