% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{failed_downloads}
\alias{failed_downloads}
\title{Failed GDC Downloads}
\format{a character vector}
\usage{
failed_downloads
}
\description{
A list of file names where the download from GDC failed
}
\keyword{datasets}
