% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{n_unique}
\alias{n_unique}
\title{Number of unique values}
\usage{
n_unique(x)
}
\arguments{
\item{x}{vector}
}
\value{
integer
}
\description{
Returns the number of unique values in a vector
}
\examples{
a <- c(1,1,2,3,4)
n_unique(a)

}
