% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_annovar.R
\name{parse_annovar}
\alias{parse_annovar}
\title{Parse ANNOVAR files}
\usage{
parse_annovar(fname)
}
\arguments{
\item{fname}{ANNOVAR file name}
}
\value{
a tibble of the results. IF the file was empty, an empty tibble
    was returned
}
\description{
Reads and parses an ANNOVAR file
}
