% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_annovar.R
\name{parse_annovar_mutation}
\alias{parse_annovar_mutation}
\title{Parse ANNOVAR mutation}
\usage{
parse_annovar_mutation(anno_tib)
}
\arguments{
\item{anno_tib}{tibble of the ANNOVAR file}
}
\value{
Returns the tibble, checking if it is empty.
    If the file was empty, an empty tibble is returned. A column is added
    that has a simple version of the amino acid change \code{aa_mod}.
}
\description{
Reads and parses an ANNOVAR file to return the amino acid change in a more
    understandable way in the column named \code{aa_mod}.
}
\examples{
extdata_dir <- system.file("extdata", package = "KrasAlleleCna")
annovar_file <- readRDS(file.path(extdata_dir, "annovar_output_list.rds"))[[3]]
annovar_file

parsed_annovar <- parse_annovar_mutation(annovar_file)
parsed_annovar

# the new column has easy to use amino acid mutations
parsed_annovar$aa_mod


}
