% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_genotype.R
\name{plot_genotype_heatmap}
\alias{plot_genotype_heatmap}
\title{Plot genotype heatmaps}
\usage{
plot_genotype_heatmap(tib, allele)
}
\arguments{
\item{tib}{copy number information}

\item{allele}{allele to plot for}
}
\description{
Plot the genotypes of mutant samples by allele as a heatmap.
    See vignette "D.01 - Plot Allele-specific Copy Number Results" for how
    to use this function. It was made into a function primarily so it could
    also be used in the README.
}
