% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_manifest}
\alias{tcga_manifest}
\title{TCGA Manifest}
\format{a tibble (1488 x 5)
\describe{
  \item{id}{file ID}
  \item{filename}{filen ame}
  \item{md5}{MD5 value for checking download success}
  \item{size}{file size (bytes(?))}
  \item{state}{internal TCGA information}
}}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
tcga_manifest
}
\description{
Manifest file downloaded from the GDC Data Portal. It contains all raw
    sequencing files available for TCGA-COAD, READ, PAAD, and DLBC, and
    TARGET-AML (not all samples were available from TARGET)
}
\keyword{datasets}
