% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_sample_sheet}
\alias{tcga_sample_sheet}
\title{TCGA Sample Sheet}
\format{a tibble (1488 x 5)
\describe{
    \item{File ID}{file ID}
    \item{File Name}{file name}
    \item{Data Category}{data category (eg. reads, mutations, etc.)}
    \item{Data Type}{data type}
    \item{Project ID}{ID of the source project (TCGA-COAD or -READ)}
    \item{Case ID}{case ID}
}}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
tcga_sample_sheet
}
\description{
Contains the information of the files in the manifest.
}
\keyword{datasets}
