% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_tcga_sample_sheet}
\alias{tidy_tcga_sample_sheet}
\title{TCGA Sample Sheet}
\format{a tibble (1488 x 5)
\describe{
    \item{file_id}{file ID}
    \item{file_name}{file name}
    \item{data_category}{category of data (are all "Sequencing Reads")}
    \item{data_type}{type of data (are all "Aligned Reads")}
    \item{project_id}{ID of the source project (TCGA-COAD or -READ)}
    \item{case_id}{case ID}
}}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
tidy_tcga_sample_sheet
}
\description{
Contains the information of the files in the manifest. The column names
    were changed to a tidy format (lowercase and spaces)
}
\keyword{datasets}
