% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{two_filename_handler}
\alias{two_filename_handler}
\alias{three_filename_handler}
\title{Select the best of two or three filenames}
\usage{
two_filename_handler(fn)

three_filename_handler(fn)
}
\arguments{
\item{fn}{a vector of filenames}
}
\value{
hopefully a single filename, but possibly more -- a message
     is printed if more than one
}
\description{
These two functions parse two or three file names based on
    my heuristics for which is most likely to be more useful.
}
