% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse a VCF file}
\usage{
parse_vcf(fname)
}
\arguments{
\item{fname}{path to the file}
}
\value{
a tibble
}
\description{
Parse the read-depth information of a VCF file.
}
\examples{
# TODO

}
